moved most complex methods into DatastoreDirectory object
- fileByName method now checks if a fileExists and if it doesn't it creates it, otherwise it retrieves
- renameFile now throws an exception if there is no file to rename
- delete now throws an exception if there is no file to delete

+++ src/main/scala/quotidian/search/DatastoreDirectory.scala
@@ -1,27 +1,48 @@
 package quotidian.search
 
-import DatastoreDirectory.{datastore,save,fileByName}
+import DatastoreDirectory.fileByName
 import com.google.appengine.api.datastore.{DatastoreService,DatastoreServiceFactory,Entity,PreparedQuery,Query}
 import com.google.appengine.api.datastore.FetchOptions.Builder.{withChunkSize}
-import java.io.Serializable
+import java.io.{IOException,Serializable}
 import java.util.Calendar
 import org.apache.lucene.store.{Directory,IndexInput,IndexOutput}
 
 class DatastoreDirectory extends Directory {
-	def close:Unit = { }
-	def createOutput(name:String):IndexOutput = null
-	def deleteFile(name:String):Unit = {
+	def close:Unit = { /* nothing to do here */ }
+	def createOutput(name:String):IndexOutput = new DatastoreIndexOutput(fileByName(name))
+	def deleteFile(name:String):Unit = DatastoreDirectory.delete(name)
+	def fileExists(name:String):Boolean = DatastoreDirectory.fileExists(name)
+	def fileLength(name:String):Long = fileByName(name).length
+	def fileModified(name:String):Long = fileByName(name).dateModified
+	def list:Array[String] = DatastoreDirectory.listFiles
+	def openInput(name:String):IndexInput = new DatastoreIndexInput(fileByName(name))
+	def renameFile(from:String,to:String):Unit = {
+		if (!DatastoreDirectory.fileExists(from)) throw new IOException("No file called {from} exists")
+		val file = fileByName(from)
+		DatastoreDirectory.save(DatastoreFile.rename(file,to))
+	}
+	def touchFile(name:String):Unit = DatastoreDirectory.save(fileByName(name))
+}
+
+object DatastoreDirectory {
+	val datastore = DatastoreServiceFactory.getDatastoreService()
+	def delete(name:String):Unit = {
+		if (!fileExists(name)) throw new IOException("No file called {name} exists")
 		val file = fileByName(name)
 		datastore.delete(file.entity.getKey)
+	}
+	def fileByName(name:String):DatastoreFile = {
+		val query = datastore.prepare(
+			new Query(DatastoreFile.Kind).addFilter(DatastoreFile.Filename,Query.FilterOperator.EQUAL,name))
+		if (fileExists(name)) DatastoreFile(query.asSingleEntity)
+		else DatastoreFile(name)
 	}
 	def fileExists(name:String):Boolean = {
 		val query = datastore.prepare(
 			new Query(DatastoreFile.Kind).addFilter(DatastoreFile.Filename,Query.FilterOperator.EQUAL,name))
 		query.countEntities > 0
 	}
-	def fileLength(name:String):Long = fileByName(name).length
-	def fileModified(name:String):Long = fileByName(name).dateModified
-	def list:Array[String] = {
+	def listFiles:Array[String] = {
 		val query = datastore.prepare(new Query(DatastoreFile.Kind))
 		val entities = query.asIterator(withChunkSize(Integer.MAX_VALUE))
 		var names = List[String]()
@@ -30,28 +51,9 @@ class DatastoreDirectory extends Directory {
 		}
 		names.toArray
 	}
-	def openInput(name:String):IndexInput = null
-	def renameFile(from:String,to:String):Unit = {
-		val file = fileByName(from)
-		save(DatastoreFile.rename(file,to))
-	}
-	def touchFile(name:String):Unit = save(fileByName(name))
-}
-
-object DatastoreDirectory {
-	val datastore = DatastoreServiceFactory.getDatastoreService()
 	def save(file:DatastoreFile):Serializable = {
 		val date = Calendar.getInstance.getTimeInMillis
 		datastore.put(file.set(DatastoreFile.DateModified,date).entity)
-	}
-	def fileByName(name:String):DatastoreFile = {
-		val query = datastore.prepare(
-			new Query(DatastoreFile.Kind).addFilter(DatastoreFile.Filename,Query.FilterOperator.EQUAL,name))
-		try {
-			DatastoreFile(query.asSingleEntity)
-		} catch {
-			case tmre:PreparedQuery.TooManyResultsException => null
-		}
 	}
 }
